requirements.txt
numpy==1.24.2
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.2.0
Pillow==9.3.0
flask==2.2.2